Simplified helpers in UK code
Since the simplified error type, brackets are needed anyway in most cases of `shouldBe` and `shouldNotBe`, and the code is mode consistent and less confusing if we use explicit list brackets rather than ordinary brackets which are magically turned into lists inside the helper.

I think the code is cleaner and clearer as a result. No change in functionality, all tests still pass.